requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
click==8.0.1
colorama==0.4.4
distlib==0.3.2
Django==3.2.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-extra-views==0.14.0
django-location-field==2.1.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
filelock==3.0.12
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
phonenumbers==8.12.30
Pillow==8.3.1
platformdirs==2.2.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
virtualenv==20.7.2
Werkzeug==2.0.1
